refactor(web): extract Markdown feature boundary
