+++ examples/web/markdown.html
@@ -352,6 +352,6 @@
     </div>
   </div>
 
-  <script type="module" src="/src/markdown-editor.ts"></script>
+  <script type="module" src="/src/entries/markdown.ts"></script>
 </body>
 </html>
