+++ examples/web/src/entries/markdown.ts
@@ -1,0 +1,3 @@
+import { mountMarkdown } from '../features/markdown/browser/mount';
+
+mountMarkdown();
