+++ examples/web/src/features/markdown/browser/mount.ts
@@ -1,0 +1,5 @@
+import { mountMarkdownApp } from './app';
+
+export function mountMarkdown(): void {
+  mountMarkdownApp();
+}
